# src/main/java/com/kshrd/assessment/docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Assessment Management API",
    "description": "API for managing assessments, exams, sections, questions, and student assessments",
    "version": "v4"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/v4/exams/{examId}": {
      "get": {
        "tags": [
          "Exam Management"
        ],
        "summary": "Get exam by ID",
        "description": "Retrieves a specific exam/assessment by its unique identifier",
        "operationId": "getExam",
        "parameters": [
          {
            "name": "examId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseExamResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Exam Management"
        ],
        "summary": "Update exam",
        "description": "Updates the details of an existing exam/assessment including name, quiz status, subject, and schedule",
        "operationId": "updateExam",
        "parameters": [
          {
            "name": "examId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExamRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseExamResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Exam Management"
        ],
        "summary": "Delete exam",
        "description": "Deletes an exam/assessment and all its associated sections and questions",
        "operationId": "deleteExam",
        "parameters": [
          {
            "name": "examId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v4/exams/{examId}/schedule": {
      "get": {
        "tags": [
          "Exam Management"
        ],
        "summary": "Get exam schedule",
        "description": "Retrieves the schedule information for a specific exam including assessment date, start time, and end time",
        "operationId": "getExamSchedule",
        "parameters": [
          {
            "name": "examId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseExamScheduleResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Exam Management"
        ],
        "summary": "Update exam schedule",
        "description": "Updates the schedule (date, start time, end time) of an exam. Can only be updated before the exam start date/time",
        "operationId": "updateSchedule",
        "parameters": [
          {
            "name": "examId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExamScheduleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseExamScheduleResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v4/exams": {
      "get": {
        "tags": [
          "Exam Management"
        ],
        "summary": "Get all exams",
        "description": "Retrieves a list of all exams/assessments in the system with pagination and search support",
        "operationId": "getAllExams",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortDirection",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "ASC"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponseExamResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Exam Management"
        ],
        "summary": "Create a new exam",
        "description": "Creates a new exam/assessment with optional sections and questions. All sections and questions can be included in the request body.",
        "operationId": "createExam",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExamRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseExamResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v4/exams/{examId}/clone": {
      "post": {
        "tags": [
          "Exam Management"
        ],
        "summary": "Clone exam",
        "description": "Creates a copy of an existing exam including all its sections and questions. The cloned exam will have '(Copy)' appended to its name",
        "operationId": "cloneExam",
        "parameters": [
          {
            "name": "examId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseExamResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v4/exams/my-exams": {
      "get": {
        "tags": [
          "Exam Management"
        ],
        "summary": "Get all exams by its owner",
        "description": "Retrieves all exams/assessments by its owner with pagination and search support",
        "operationId": "getMyExams",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortDirection",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "ASC"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponseExamResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v4/exams/active": {
      "get": {
        "tags": [
          "Exam Management"
        ],
        "summary": "Get active exams",
        "description": "Retrieves a list of currently active exams with pagination and search support. An exam is active if the current time is within the scheduled start and end time on the assessment date.",
        "operationId": "getActiveExams",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortDirection",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "ASC"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponseExamResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponseVoid": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {

          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "errorCode": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        }
      },
...
      "ExamRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "isQuiz": {
            "type": "boolean"
          },
          "subjectId": {
            "type": "string",
            "format": "uuid"
          },
          "schedule": {
            "$ref": "#/components/schemas/Schedule"
          },
          "sections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SectionRequest"
            }
          }
        },
        "required": [
          "isQuiz",
          "name",
          "subjectId"
        ]
      },
      "QuestionRequest": {
        "type": "object",
        "properties": {
          "questionType": {
            "type": "string",
            "enum": [
              "Multiple Choice",
              "True/False",
              "Long Answer",
              "Code Problem"
            ]
          },
          "image": {
            "type": "string"
          },
          "questionContent": {
            "type": "object",
            "additionalProperties": {

            }
          },
          "points": {
            "type": "number",
            "format": "double"
          }
        },
        "required": [
          "questionContent",
          "questionType"
        ]
      },
      "Schedule": {
        "type": "object",
        "properties": {
          "assessmentDate": {
            "type": "string",
            "format": "date"
          },
          "startTime": {
            "type": "string"
          },
          "endTime": {
            "type": "string"
          }
        },
        "required": [
          "assessmentDate",
          "endTime",
          "startTime"
        ]
      },
      "SectionRequest": {
        "type": "object",
        "properties": {
          "sectionName": {
            "type": "string",
            "minLength": 1
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionRequest"
            }
          }
        },
        "required": [
          "sectionName"
        ]
      },
      "ApiResponseExamResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/ExamResponse"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "errorCode": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        }
      },
      "ExamResponse": {
        "type": "object",
        "properties": {
          "assessmentId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "isQuiz": {
            "type": "boolean"
          },
          "subjectId": {
            "type": "string",
            "format": "uuid"
          },
          "schedule": {
            "$ref": "#/components/schemas/Schedule"
          },
          "createdBy": {
            "type": "string",
            "format": "uuid"
          },
          "updatedBy": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "totalSections": {
            "type": "integer",
            "format": "int64"
          },
          "totalQuestions": {
            "type": "integer",
            "format": "int64"
          },
          "sections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SectionResponse"
            }
          }
        }
      },
      "QuestionResponse": {
        "type": "object",
        "properties": {
          "questionId": {
            "type": "string",
            "format": "uuid"
          },
          "questionType": {
            "type": "string",
            "enum": [
              "Multiple Choice",
              "True/False",
              "Long Answer",
              "Code Problem"
            ]
          },
          "image": {
            "type": "string"
          },
          "questionContent": {
            "type": "object",
            "additionalProperties": {

            }
          },
          "sectionId": {
            "type": "string",
            "format": "uuid"
          },
          "points": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "SectionResponse": {
        "type": "object",
        "properties": {
          "sectionId": {
            "type": "string",
            "format": "uuid"
          },
          "sectionName": {
            "type": "string"
          },
          "assessmentId": {
            "type": "string",
            "format": "uuid"
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionResponse"
            }
          }
        }
      },
      "ExamScheduleRequest": {
        "type": "object",
        "properties": {
          "assessmentDate": {
            "type": "string",
            "format": "date"
          },
          "startTime": {
            "type": "string"
          },
          "endTime": {
            "type": "string"
          }
        },
        "required": [
          "assessmentDate",
          "endTime",
          "startTime"
        ]
      },
      "ApiResponseExamScheduleResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/ExamScheduleResponse"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "errorCode": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        }
      },
      "ExamScheduleResponse": {
        "type": "object",
        "properties": {
          "assessmentDate": {
            "type": "string",
            "format": "date"
          },
          "startTime": {
            "type": "string"
          },
          "endTime": {
            "type": "string"
          }
        }
      },
...
      "ApiResponsePageResponseExamResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PageResponseExamResponse"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "errorCode": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        }
      },
      "PageResponseExamResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExamResponse"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "hasNext": {
            "type": "boolean"
          },
          "hasPrevious": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}